# api/openapi.yaml
openapi: 3.0.3
info:
  title: LecodeKanban API
  version: 0.3.0
  description: |
    REST API of LecodeKanban. Source of truth for Go DTOs and the TypeScript client (`make gen`).

    **Auth:** httpOnly cookies — `lk_at` (access, ~15 min) and `lk_rt` (rotating refresh,
    path-scoped to `/api/v1/auth`). **CSRF:** every non-GET request must echo the `lk_csrf`
    cookie in the `X-CSRF-Token` header (obtain it via `GET /auth/csrf`).

    **Errors** are `{"error": {"code", "message", "fields"?, "meta"?}}`. `code` is stable and
    translated by clients; `message` is an English developer hint. See `GET /i18n/error-codes`.
servers:
  - url: /api/v1
...
paths:
...
  /cards/{cardId}/comments:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [comments]
      operationId: listComments
      responses:
        "200":
          description: Comments, oldest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Comment" } }
    post:
      tags: [comments]
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "422": { $ref: "#/components/responses/Error" }
  /comments/{commentId}:
    parameters: [{ $ref: "#/components/parameters/CommentId" }]
    patch:
      tags: [comments]
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "403": { $ref: "#/components/responses/Error" }
    delete:
      tags: [comments]
      operationId: deleteComment
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lk_at }
  parameters:
...
    CardId:
      name: cardId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CommentId: { name: commentId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: auth.invalid_credentials }
            message: { type: string }
            fields:
              type: array
              items: { $ref: "#/components/schemas/FieldError" }
            meta:
              type: object
              additionalProperties: true
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, example: min_length }
        params: { type: object, additionalProperties: true }
...
    PersonRef:
      type: object
      required: [id, name, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
...
    Comment:
      type: object
      required: [id, cardId, author, body, mentions, createdAt, editedAt]
      properties:
        id: { type: string, format: uuid }
        cardId: { type: string, format: uuid }
        author: { allOf: [{ $ref: "#/components/schemas/PersonRef" }], nullable: true }
        body: { type: string, description: "Markdown; mentions are written as @[Name](user-id)" }
        mentions: { type: array, items: { $ref: "#/components/schemas/PersonRef" } }
        createdAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
    CommentInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 10000 }
...
security:
  - cookieAuth: []